requirements.txt
# python -m venv venv
# venv\Scripts\activate
# python -m pip install --upgrade pip
# pip install -r requirements.txt
# pip install flash_attn-2.8.0.post2-cp310-cp310-win_amd64.whl

--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
transformers
accelerate
pillow
bitsandbytes
triton-windows<3.4
flask